Change lut_vector_indexes type

+++ concrete-cuda/cuda/src/bit_extraction.cu
@@ -184,7 +184,7 @@ void cuda_extract_bits_64(
         (uint64_t *)lwe_array_in_shifted_buffer,
         (uint64_t *)lwe_array_out_ks_buffer,
         (uint64_t *)lwe_array_out_pbs_buffer, (uint64_t *)lut_pbs,
-        (uint32_t *)lut_vector_indexes, (uint64_t *)ksk, (double2 *)fourier_bsk,
+        (uint64_t *)lut_vector_indexes, (uint64_t *)ksk, (double2 *)fourier_bsk,
         number_of_bits, delta_log, lwe_dimension_in, lwe_dimension_out,
         base_log_bsk, level_count_bsk, base_log_ksk, level_count_ksk,
         number_of_samples, max_shared_memory);
@@ -196,7 +196,7 @@ void cuda_extract_bits_64(
         (uint64_t *)lwe_array_in_shifted_buffer,
         (uint64_t *)lwe_array_out_ks_buffer,
         (uint64_t *)lwe_array_out_pbs_buffer, (uint64_t *)lut_pbs,
-        (uint32_t *)lut_vector_indexes, (uint64_t *)ksk, (double2 *)fourier_bsk,
+        (uint64_t *)lut_vector_indexes, (uint64_t *)ksk, (double2 *)fourier_bsk,
         number_of_bits, delta_log, lwe_dimension_in, lwe_dimension_out,
         base_log_bsk, level_count_bsk, base_log_ksk, level_count_ksk,
         number_of_samples, max_shared_memory);
@@ -208,7 +208,7 @@ void cuda_extract_bits_64(
         (uint64_t *)lwe_array_in_shifted_buffer,
         (uint64_t *)lwe_array_out_ks_buffer,
         (uint64_t *)lwe_array_out_pbs_buffer, (uint64_t *)lut_pbs,
-        (uint32_t *)lut_vector_indexes, (uint64_t *)ksk, (double2 *)fourier_bsk,
+        (uint64_t *)lut_vector_indexes, (uint64_t *)ksk, (double2 *)fourier_bsk,
         number_of_bits, delta_log, lwe_dimension_in, lwe_dimension_out,
         base_log_bsk, level_count_bsk, base_log_ksk, level_count_ksk,
         number_of_samples, max_shared_memory);
@@ -220,7 +220,7 @@ void cuda_extract_bits_64(
         (uint64_t *)lwe_array_in_shifted_buffer,
         (uint64_t *)lwe_array_out_ks_buffer,
         (uint64_t *)lwe_array_out_pbs_buffer, (uint64_t *)lut_pbs,
-        (uint32_t *)lut_vector_indexes, (uint64_t *)ksk, (double2 *)fourier_bsk,
+        (uint64_t *)lut_vector_indexes, (uint64_t *)ksk, (double2 *)fourier_bsk,
         number_of_bits, delta_log, lwe_dimension_in, lwe_dimension_out,
         base_log_bsk, level_count_bsk, base_log_ksk, level_count_ksk,
         number_of_samples, max_shared_memory);
@@ -232,7 +232,7 @@ void cuda_extract_bits_64(
         (uint64_t *)lwe_array_in_shifted_buffer,
         (uint64_t *)lwe_array_out_ks_buffer,
         (uint64_t *)lwe_array_out_pbs_buffer, (uint64_t *)lut_pbs,
-        (uint32_t *)lut_vector_indexes, (uint64_t *)ksk, (double2 *)fourier_bsk,
+        (uint64_t *)lut_vector_indexes, (uint64_t *)ksk, (double2 *)fourier_bsk,
         number_of_bits, delta_log, lwe_dimension_in, lwe_dimension_out,
         base_log_bsk, level_count_bsk, base_log_ksk, level_count_ksk,
         number_of_samples, max_shared_memory);
